commit hbase code

+++ hbase_test/src/main/java/com/lendou/hbase/HBaseHelper.java
@@ -4,19 +4,15 @@ import org.apache.commons.lang.Validate;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.*;
 import org.apache.hadoop.hbase.client.*;
-import org.apache.hadoop.hbase.util.Bytes;
 
 import java.io.IOException;
-import java.util.Iterator;
+import java.util.List;
 
 /**
- * Hello world!
- *
+ * HBase的帮助类, 用于封装访问hbase的基本方法
  */
-public class Program {
+public class HBaseHelper {
     private static Configuration conf = null;
-
-    private static final String TABLE_NAME = "test";
 
     /**
      * Initialization
@@ -25,6 +21,12 @@ public class Program {
         conf = HBaseConfiguration.create();
     }
 
+    /**
+     * 创建HBase表
+     *
+     * @param tableName 待创建的表名
+     * @param families  待创建表的family名称
+     */
     public static void createTable(String tableName, String[] families) {
         Validate.notNull(tableName);
         Validate.notNull(families);
@@ -34,16 +36,16 @@ public class Program {
         try {
             conn = ConnectionFactory.createConnection(conf);
             admin = (HBaseAdmin) conn.getAdmin();
-            if (admin.tableExists(TABLE_NAME)) {
-                System.out.println("Table " + TABLE_NAME + " already exists!");
-
+            if (admin.tableExists(tableName)) {
+                if (!admin.isTableEnabled(tableName)) {
+                    admin.enableTable(tableName);
+                }
             } else {
-                HTableDescriptor tableDesc = new HTableDescriptor(TableName.valueOf(TABLE_NAME));
+                HTableDescriptor tableDesc = new HTableDescriptor(TableName.valueOf(tableName));
                 for (int i = 0; i < families.length; i++) {
                     tableDesc.addFamily(new HColumnDescriptor(families[i]));
                 }
                 admin.createTable(tableDesc);
-                System.out.println("Create table " + tableName + " ok.");
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -66,6 +68,11 @@ public class Program {
         }
     }
 
+    /**
+     * 删除HBase表
+     *
+     * @param tableName 待删除的HBase表名称
+     */
     public static void deleteTable(String tableName) {
         Validate.notEmpty(tableName);
 
@@ -74,7 +81,14 @@ public class Program {
         try {
             conn = ConnectionFactory.createConnection(conf);
             admin = (HBaseAdmin) conn.getAdmin();
-            admin.disableTable(tableName);
+            if (!admin.tableExists(tableName)) {
+                return;
+            }
+
+            if (admin.isTableEnabled(tableName)) {
+                admin.disableTable(tableName);
+            }
+
             admin.deleteTable(tableName);
         } catch (IOException e) {
             e.printStackTrace();
@@ -97,23 +111,22 @@ public class Program {
         }
     }
 
-    public static void addRecord(String tableName, String rowKey, String family, String qualifier, String value) {
+    /**
+     * 写入一个数据
+     *
+     * @param tableName 待写入数据的表名称
+     * @param put       要写入的部分
+     */
+    public static void put(String tableName, Put put) {
         Validate.notEmpty(tableName);
-        Validate.notEmpty(rowKey);
-        Validate.notEmpty(family);
-        Validate.notEmpty(qualifier);
-        Validate.notEmpty(value);
+        Validate.notNull(put);
 
         Connection conn = null;
         HTable table = null;
         try {
             conn = ConnectionFactory.createConnection(conf);
             table = (HTable) conn.getTable(TableName.valueOf(tableName));
-
-            Put put = new Put(Bytes.toBytes(rowKey));
-            put.addColumn(Bytes.toBytes(family), Bytes.toBytes(qualifier), Bytes.toBytes(value));
             table.put(put);
-            System.out.println("insert record " + rowKey + " to table " + tableName + " ok.");
 
         } catch (IOException e) {
             e.printStackTrace();
@@ -136,20 +149,59 @@ public class Program {
         }
     }
 
-    public static void deleteRecord(String tableName, String rowKey) {
+    /**
+     * 批量写入一组数据
+     *
+     * @param tableName 待写入数据的表名称
+     * @param puts      批量写入的数据
+     */
+    public static void put(String tableName, List<Put> puts) {
         Validate.notEmpty(tableName);
-        Validate.notEmpty(rowKey);
+        Validate.notNull(puts);
 
         Connection conn = null;
         HTable table = null;
         try {
             conn = ConnectionFactory.createConnection(conf);
             table = (HTable) conn.getTable(TableName.valueOf(tableName));
+            table.put(puts);
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (table != null) {
+                try {
+                    table.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
 
-            Delete delete = new Delete(Bytes.toBytes(rowKey));
+            if (conn != null) {
+                try {
+                    conn.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+    }
+
+    /**
+     * 删除一个数据
+     *
+     * @param tableName 待删除数据的表名称
+     * @param delete    带删除的数据
+     */
+    public static void delete(String tableName, Delete delete) {
+        Validate.notEmpty(tableName);
+        Validate.notNull(delete);
+
+        Connection conn = null;
+        HTable table = null;
+        try {
+            conn = ConnectionFactory.createConnection(conf);
+            table = (HTable) conn.getTable(TableName.valueOf(tableName));
             table.delete(delete);
-            System.out.println("del record " + rowKey + " ok.");
-
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -171,21 +223,22 @@ public class Program {
         }
     }
 
-    public static void deleteRecord(String tableName, String rowKey, String family, String qualifier) {
+    /**
+     * 批量删除一组数据
+     *
+     * @param tableName 待删除数据的表名称
+     * @param deletes   待删除的数据
+     */
+    public static void delete(String tableName, List<Delete> deletes) {
         Validate.notEmpty(tableName);
-        Validate.notEmpty(rowKey);
+        Validate.notNull(deletes);
 
         Connection conn = null;
         HTable table = null;
         try {
             conn = ConnectionFactory.createConnection(conf);
             table = (HTable) conn.getTable(TableName.valueOf(tableName));
-
-            Delete delete = new Delete(Bytes.toBytes(rowKey));
-            delete.addColumn(Bytes.toBytes(family), Bytes.toBytes(qualifier));
-            table.delete(delete);
-            System.out.println("del record " + rowKey + " ok.");
-
+            table.delete(deletes);
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -207,26 +260,34 @@ public class Program {
         }
     }
 
-    public static void getRecord(String tableName, String rowKey) {
+    /**
+     * 查询一个数据
+     *
+     * @param tableName 待查询数据的表名称
+     * @param get       待查询的数据
+     * @return          查询结果
+     */
+    public static Result get(String tableName, Get get) {
         Validate.notEmpty(tableName);
-        Validate.notEmpty(rowKey);
+        Validate.notNull(get);
 
         Connection conn = null;
         HTable table = null;
         try {
             conn = ConnectionFactory.createConnection(conf);
             table = (HTable) conn.getTable(TableName.valueOf(tableName));
-
-            Get get = new Get(Bytes.toBytes(rowKey));
-            Result rs = table.get(get);
-            for (Cell cell : rs.rawCells()) {
-                String rowValue = Bytes.toString(Bytes.copy(cell.getRowArray(), cell.getRowOffset(), cell.getRowLength()));
-                String familyValue = Bytes.toString(Bytes.copy(cell.getFamilyArray(), cell.getFamilyOffset(), cell.getFamilyLength()));
-                String qualifierValue = Bytes.toString(Bytes.copy(cell.getQualifierArray(), cell.getQualifierOffset(), cell.getQualifierLength()));
-                String value = Bytes.toString(Bytes.copy(cell.getValueArray(), cell.getValueOffset(), cell.getValueLength()));
-
-                System.out.println(String.format("%s\t%s:%s\t%s", rowValue, familyValue, qualifierValue, value));
-            }
+            return table.get(get);
+//
+//            Get get = new Get(Bytes.toBytes(rowKey));
+//            Result rs = table.get(get);
+//            for (Cell cell : rs.rawCells()) {
+//                String rowValue = Bytes.toString(Bytes.copy(cell.getRowArray(), cell.getRowOffset(), cell.getRowLength()));
+//                String familyValue = Bytes.toString(Bytes.copy(cell.getFamilyArray(), cell.getFamilyOffset(), cell.getFamilyLength()));
+//                String qualifierValue = Bytes.toString(Bytes.copy(cell.getQualifierArray(), cell.getQualifierOffset(), cell.getQualifierLength()));
+//                String value = Bytes.toString(Bytes.copy(cell.getValueArray(), cell.getValueOffset(), cell.getValueLength()));
+//
+//                System.out.println(String.format("%s\t%s:%s\t%s", rowValue, familyValue, qualifierValue, value));
+//            }
 
         } catch (IOException e) {
             e.printStackTrace();
@@ -247,32 +308,27 @@ public class Program {
                 }
             }
         }
+
+        return null;
     }
 
-    public static void getRecords(String tableName) {
+    /**
+     * 批量查询一组数据
+     *
+     * @param tableName 待查询数据的表名称
+     * @param gets      需要查询的数据
+     * @return          查询结果
+     */
+    public static Result[] get(String tableName, List<Get> gets) {
         Validate.notEmpty(tableName);
+        Validate.notNull(gets);
 
         Connection conn = null;
         HTable table = null;
         try {
             conn = ConnectionFactory.createConnection(conf);
             table = (HTable) conn.getTable(TableName.valueOf(tableName));
-
-            Scan scan = new Scan();
-            ResultScanner rss = table.getScanner(scan);
-            Iterator<Result> iter = rss.iterator();
-            while (iter.hasNext()) {
-                Result rs = iter.next();
-                for (Cell cell : rs.rawCells()) {
-                    String rowValue = Bytes.toString(Bytes.copy(cell.getRowArray(), cell.getRowOffset(), cell.getRowLength()));
-                    String familyValue = Bytes.toString(Bytes.copy(cell.getFamilyArray(), cell.getFamilyOffset(), cell.getFamilyLength()));
-                    String qualifierValue = Bytes.toString(Bytes.copy(cell.getQualifierArray(), cell.getQualifierOffset(), cell.getQualifierLength()));
-                    String value = Bytes.toString(Bytes.copy(cell.getValueArray(), cell.getValueOffset(), cell.getValueLength()));
-
-                    System.out.println(String.format("%s\t%s:%s\t%s\t%d", rowValue, familyValue, qualifierValue, value, cell.getTimestamp()));
-                }
-            }
-
+            return table.get(gets);
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -292,10 +348,47 @@ public class Program {
                 }
             }
         }
+
+        return null;
     }
 
-    public static void main( String[] args ) throws IOException {
-        addRecord(TABLE_NAME, "r1", "x", "a", "value-a1");
-        getRecords(TABLE_NAME);
+    /**
+     * 扫描一组数据
+     *
+     * @param tableName 待扫描数据的表名称
+     * @param scan      扫描数据所使用的scan结构
+     * @return          扫描结果
+     */
+    public static ResultScanner scan(String tableName, Scan scan) {
+        Validate.notEmpty(tableName);
+        Validate.notNull(scan);
+
+        Connection conn = null;
+        HTable table = null;
+        try {
+            conn = ConnectionFactory.createConnection(conf);
+            table = (HTable) conn.getTable(TableName.valueOf(tableName));
+            return table.getScanner(scan);
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (table != null) {
+                try {
+                    table.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            if (conn != null) {
+                try {
+                    conn.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+        return null;
     }
 }
